Add subscription resource - CodeTriage
For large repos like `rails/rails` subscribing to receive all GitHub notifications is overwhelming and demoralizing. I wrote a service, CodeTriage, that lets you meter out the number of issues you get in your inbox. It also provides people with prompts for things to do with the issues when they get them so they're not just left hanging. When people stop interacting with the emails it backs off to weekly, bi-weekly, and then monthly so that people have time to burn out.

I think it would be worth mentioning it as an option in the "subscribe" section. Let me know if you have any questions or comments.

Love the post btw.